+++ lua/plugins/treesitter.lua
@@ -1,0 +1,24 @@
+-- Highlight, edit, and navigate code
+return {
+  'nvim-treesitter/nvim-treesitter',
+  build = ':TSUpdate',
+  opts = {
+    ensure_installed = { 'bash', 'c', 'diff', 'html', 'javascript', 'lua', 'luadoc', 'markdown', 'python', 'vim', 'vimdoc' },
+    -- Autoinstall languages that are not installed
+    auto_install = true,
+    highlight = {
+      enable = true,
+      -- Some languages depend on vim's regex highlighting system (such as Ruby) for indent rules.
+      --  If you are experiencing weird indenting issues, add the language to
+      --  the list of additional_vim_regex_highlighting and disabled languages for indent.
+      additional_vim_regex_highlighting = { 'ruby' },
+      disable = { 'dockerfile' },
+    },
+    indent = { enable = true, disable = { 'ruby' } },
+  },
+  config = function(_, opts)
+    -- [[ Configure Treesitter ]] See `:help nvim-treesitter`
+    require('nvim-treesitter.install').prefer_git = true
+    require('nvim-treesitter.configs').setup(opts)
+  end,
+}
